backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
requests
pyttsx3
openai-whisper